+++ src/main/java/de/guntram/mcmod/durabilityviewer/config/ModMenuImpl.java
@@ -2,13 +2,15 @@ package de.guntram.mcmod.durabilityviewer.config;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
+import de.tobi.voxelconfig.ConfigScreen;
+import net.minecraft.network.chat.Component;
 
 public class ModMenuImpl implements ModMenuApi {
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return (screen) -> {
-            ConfigGui gui = new ConfigGui();
-            gui.setParent(screen);
-            return gui;
-        };
+        return (screen) -> ConfigScreen.create(
+                Component.translatable("durabilityviewer.gui.title.configs"),
+                Configs.INSTANCE,
+                screen,
+                Configs::saveToFile);
     }
 }