+++ src/main/java/de/guntram/mcmod/durabilityviewer/config/WarnMode.java
@@ -1,11 +1,8 @@
 package de.guntram.mcmod.durabilityviewer.config;
 
-import com.google.common.collect.ImmutableList;
-import fi.dy.masa.malilib.config.IConfigOptionList;
-import fi.dy.masa.malilib.config.IConfigOptionListEntry;
-import fi.dy.masa.malilib.util.StringUtils;
+import net.minecraft.network.chat.Component;
 
-public enum WarnMode implements IConfigOptionListEntry {
+public enum WarnMode {
     NONE("None", "durabilityviewer.config.warnmode.none"),
     SOUND("Sound", "durabilityviewer.config.warnmode.sound"),
     VISUAL("Visual", "durabilityviewer.config.warnmode.visual"),
@@ -19,45 +16,14 @@ public enum WarnMode implements IConfigOptionListEntry {
         this.translationKey = translationKey;
     }
 
-    @Override
     public String getStringValue()
     {
         return this.configString;
     }
 
-    @Override
-    public String getDisplayName()
+    public Component getDisplayName()
     {
-        return StringUtils.translate(this.translationKey);
-    }
-
-    @Override
-    public IConfigOptionListEntry cycle(boolean forward)
-    {
-        int id = this.ordinal();
-
-        if (forward)
-        {
-            if (++id >= values().length)
-            {
-                id = 0;
-            }
-        }
-        else
-        {
-            if (--id < 0)
-            {
-                id = values().length - 1;
-            }
-        }
-
-        return values()[id % values().length];
-    }
-
-    @Override
-    public WarnMode fromString(String name)
-    {
-        return fromStringStatic(name);
+        return Component.translatable(this.translationKey);
     }
 
     public static WarnMode fromStringStatic(String name)